Moved the machine code into `code.h`
... and a few more interesting things. Now the unroll count is
arbitrary, automatically stopping execution once the function
returns. Assuming the given function converges and returns it should
work perfectly for any function.

Minimal assumptions are a problem: need to assume that one could copy
the machine code, that sp does not overflow and others... Potentially
this could be pruned on the basis of "move the code to another
location and see if it works"? Or hoare - "this assumes the linker
puts it somewhere in this region"?